+++ gateways/db/mysql/user/list.go
@@ -2,6 +2,7 @@ package user
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/TLSDevv/golang_catatan_keuangan_backend/domain/entities"
 )
@@ -9,7 +10,7 @@ import (
 func (u UserRepository) List(ctx context.Context) ([]entities.User, error) {
 	sql := `
 		SELECT
-			id, username, email, password, fullname, role, created_at, updated_at,
+			id, username, email, password, fullname, role, created_at, updated_at
 		FROM
 			users
 		WHERE
@@ -43,6 +44,8 @@ func (u UserRepository) List(ctx context.Context) ([]entities.User, error) {
 		}
 		users = append(users, user)
 	}
+
+	fmt.Println(err)
 
 	return users, nil
 }
